Doctrine ORM first implementation and testing done

+++ src/Entity/Product.php
@@ -17,7 +17,7 @@ class Product
     private ?string $name = null;
 
     #[ORM\Column]
-    private ?int $price = null;
+    private ?float $price = null;
 
     public function getId(): ?int
     {
@@ -36,12 +36,12 @@ class Product
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
 
-    public function setPrice(int $price): static
+    public function setPrice(float $price): static
     {
         $this->price = $price;
